Рефакторинг кода: переработана структура приложения
