+++ data/data.json
@@ -1,0 +1,1 @@
+{"Visa invitation": "There is nothing to worry about. The invitation is issued by the Ministry of Foreign Affairs of the Russian Federation and takes at least 30 days to be prepared. If you do not receive your invitation 40 days after submitting the application, please contact the Admissions Office at admission@urfu.ru.", "Visa extension": "Please visit the GUK-109 office at least six weeks prior to the expiration date of your current visa. You will be consulted on the visa extension procedures.", "Accommodation for international students": "All international students of UrFU receive a place in one of the campus dormitories. The approximate accommodation fee is 1800 rubles per month. You can find more info here.", "Amount of people in a dormitory room": "There are usually two or three people living in one dormitory room.", "Dormitory room readiness": "Every student has the right to get a room in an university dormitory if places are available or free. However, please keep in mind that, according to the legislation, every student must undergo a standard medical  examination before they are allowed to check into a university dormitory room.  Since the medical examination may take up more than one day, you might need to book a hotel room for this time.  Should you need any help with booking, please contact our Buddy System coordinators.", "Dormitory address": "The address of your dormitory will be written in the voucher for the dormitory, which you will receive in the International Students Adaptation Center.", "Check in": "According to the legislation, all students must undergo a standard medical examination before checking into a dormitory room. You will receive all required information in the Admissions Office (19 Mira Street, office GUK-109).", "Other accommodation options": "If you prefer renting a flat to living in a dormitory, our Student Union can help you find a suitable paid accommodation and \u2013 if necessary \u2013 a flatmate.", "Russian higher education system": "The Russian education system is a multistage process that encourages lifelong learning. This system includes compulsory general education (kindergartens and schools) and professional education (colleges, technical schools, universities). In turn, professional education is divided into secondary and higher education. Professional development and retraining programs are considered the final educational stage.", "Point-Grade System": "The Point-Grade System is a digital mark for evaluating students' learning activities and achievements. The Point-Grade System is a system for determining the level of student success based on the cumulative principle of assessing academic performance and its results.", "Lectures begin and finish": "All the information about dates can be found here: https://urfu.ru/en/current-students/academic-planning/academic-calendar/", "Not identified": "Cannot find any information on this topic. Please rephrase your question!"}
